Separate out the net symbols from standard linux
Also includes some typing changes.
