+++ volatility3/framework/plugins/linux/netfilter.py
@@ -17,6 +17,7 @@ from volatility3.framework import (
 from volatility3.framework.renderers import format_hints
 from volatility3.framework.configuration import requirements
 from volatility3.framework.symbols import linux
+from volatility3.framework.symbols.linux import net
 from volatility3.plugins.linux import lsmod
 
 vollog = logging.getLogger(__name__)
@@ -82,7 +83,7 @@ class AbstractNetfilter(ABC):
         self.list_head_size = self.vmlinux.get_type("list_head").size
 
         lsmod_required_version = Netfilter._required_lsmod_version
-        lsmod_current_version = lsmod.Lsmod._version
+        lsmod_current_version = lsmod.Lsmod.version
         if not requirements.VersionRequirement.matches_required(
             lsmod_required_version, lsmod_current_version
         ):
@@ -91,7 +92,7 @@ class AbstractNetfilter(ABC):
             )
 
         linuxutils_required_version = Netfilter._required_linuxutils_version
-        linuxutils_current_version = linux.LinuxUtilities._version
+        linuxutils_current_version = linux.LinuxUtilities.version
         if not requirements.VersionRequirement.matches_required(
             linuxutils_required_version, linuxutils_current_version
         ):
@@ -99,11 +100,20 @@ class AbstractNetfilter(ABC):
                 f"linux.LinuxUtilities version not suitable: required {linuxutils_required_version} found {linuxutils_current_version}"
             )
 
+        linux_net_required_version = Netfilter._required_linuxnet_version
+        linux_net_current_version = net.NetSymbols.version
+        if not requirements.VersionRequirement.matches_required(
+            linux_net_required_version, linux_net_current_version
+        ):
+            raise exceptions.PluginRequirementException(
+                f"symbols.linux.net.NetSymbols version not suitable: required {linux_net_required_version} found {linux_net_current_version}"
+            )
+
         linux_utilities_modules_required_version = (
             Netfilter._required_linux_utilities_modules_version
         )
         linux_utilities_modules_current_version = (
-            linux_utilities_modules.Modules._version
+            linux_utilities_modules.Modules.version
         )
         if not requirements.VersionRequirement.matches_required(
             linux_utilities_modules_required_version,
@@ -112,6 +122,9 @@ class AbstractNetfilter(ABC):
             raise exceptions.PluginRequirementException(
                 f"linux_utilities_modules.Modules version not suitable: required {linux_utilities_modules_required_version} found {linux_utilities_modules_current_version}"
             )
+
+        symbol_table = self._context.symbol_space[self.vmlinux.symbol_table_name]
+        net.NetSymbols.apply(symbol_table)
 
         modules = lsmod.Lsmod.list_modules(context, kernel_module_name)
         self.handlers = linux.LinuxUtilities.generate_kernel_handler_info(
@@ -697,6 +710,7 @@ class Netfilter(interfaces.plugins.PluginInterface):
     _required_linux_utilities_modules_version = (1, 0, 0)
     _required_linuxutils_version = (2, 1, 0)
     _required_lsmod_version = (2, 0, 0)
+    _required_linuxnet_version = (1, 0, 0)
 
     @classmethod
     def get_requirements(cls) -> List[interfaces.configuration.RequirementInterface]:
@@ -718,6 +732,11 @@ class Netfilter(interfaces.plugins.PluginInterface):
                 name="linuxutils",
                 component=linux.LinuxUtilities,
                 version=cls._required_linuxutils_version,
+            ),
+            requirements.VersionRequirement(
+                name="linuxnet",
+                component=net.NetSymbols,
+                version=cls._required_linuxnet_version,
             ),
         ]
