+++ volatility3/framework/symbols/linux/__init__.py
@@ -64,30 +64,31 @@ class LinuxKernelIntermedSymbols(intermed.IntermediateSymbolTable):
         # kernels < 4.18. Reuses timespec64 obj extension, since both has the same members
         self.optional_set_type_class("timespec", extensions.timespec64)
 
+        # Network
+        # FIXME: Deprecate all of this once the framework hits version 3
+        self.set_type_class("net", extensions.net.net)
+        self.set_type_class("net_device", extensions.net.net_device)
+        self.set_type_class("in_device", extensions.net.in_device)
+        self.set_type_class("in_ifaddr", extensions.net.in_ifaddr)
+        self.set_type_class("inet6_dev", extensions.net.inet6_dev)
+        self.set_type_class("inet6_ifaddr", extensions.net.inet6_ifaddr)
+        self.set_type_class("socket", extensions.net.socket)
+        self.set_type_class("sock", extensions.net.sock)
+        self.set_type_class("inet_sock", extensions.net.inet_sock)
+        self.set_type_class("unix_sock", extensions.net.unix_sock)
+        # Might not exist in older kernels or the current symbols
+        self.optional_set_type_class("netlink_sock", extensions.net.netlink_sock)
+        self.optional_set_type_class("vsock_sock", extensions.net.vsock_sock)
+        self.optional_set_type_class("packet_sock", extensions.net.packet_sock)
+        self.optional_set_type_class("bt_sock", extensions.net.bt_sock)
+        self.optional_set_type_class("xdp_sock", extensions.net.xdp_sock)
+
         # Mount
         self.set_type_class("vfsmount", extensions.vfsmount)
         # Might not exist in older kernels or the current symbols
         self.optional_set_type_class("mount", extensions.mount)
         self.optional_set_type_class("mnt_namespace", extensions.mnt_namespace)
         self.optional_set_type_class("rb_root", extensions.rb_root)
-
-        # Network
-        self.set_type_class("net", extensions.net)
-        self.set_type_class("net_device", extensions.net_device)
-        self.set_type_class("in_device", extensions.in_device)
-        self.set_type_class("in_ifaddr", extensions.in_ifaddr)
-        self.set_type_class("inet6_dev", extensions.inet6_dev)
-        self.set_type_class("inet6_ifaddr", extensions.inet6_ifaddr)
-        self.set_type_class("socket", extensions.socket)
-        self.set_type_class("sock", extensions.sock)
-        self.set_type_class("inet_sock", extensions.inet_sock)
-        self.set_type_class("unix_sock", extensions.unix_sock)
-        # Might not exist in older kernels or the current symbols
-        self.optional_set_type_class("netlink_sock", extensions.netlink_sock)
-        self.optional_set_type_class("vsock_sock", extensions.vsock_sock)
-        self.optional_set_type_class("packet_sock", extensions.packet_sock)
-        self.optional_set_type_class("bt_sock", extensions.bt_sock)
-        self.optional_set_type_class("xdp_sock", extensions.xdp_sock)
 
         # Only found in 6.1+ kernels
         self.optional_set_type_class("maple_tree", extensions.maple_tree)
